requirements.txt
openai-agents
flask
python-dotenv
